CLEANUP: Move Breakpoint to Domain/Model folder
